update fitur app

+++ app/Http/Controllers/ApplicationController.php
@@ -18,6 +18,9 @@ use App\Models\Opd;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+
+use function PHPUnit\Framework\isEmpty;
 
 class ApplicationController extends Controller
 {
@@ -38,20 +41,15 @@ class ApplicationController extends Controller
      */
     public function create()
     {
-        $katapps = Katapp::all();
-        $katpenggunas = Katpengguna::all();
-        $katplatforms = Katplatform::all();
-        $katdbs = Katdb::all();
-        $katservers = Katserver::all();
-        $bahasaprograms = Bahasaprogram::all();
-        $frameworkapps = Frameworkapp::all();
-        $layananapps = Layananapp::all();
         $opds = Opd::all();
+        $katpenggunas = Katpengguna::all();
+        $katservers = Katserver::all();
+        $layananapps = Layananapp::all();
+        $layananapps = Layananapp::all();
+        $katapps = Katapp::whereIn('id', [1, 2])->get();
+        // $noRegis = Str::upper(Str::random(8));
 
-        // $response = Http::get('https://splp.denpasarkota.go.id/index.php/dev/master/opd');
-        // $opds = $response->json(['entry']);
-
-        return view('aplikasi.aplikasi-create', compact('katapps', 'katpenggunas', 'katplatforms', 'katdbs', 'katservers', 'bahasaprograms', 'frameworkapps', 'layananapps', 'opds'), [
+        return view('aplikasi.aplikasi-create', compact('opds', 'katpenggunas', 'katservers', 'layananapps', 'katapps'), [
             'title' => 'Data Aplikasi'
         ]);
     }
@@ -64,49 +62,11 @@ class ApplicationController extends Controller
         DB::transaction(function () use ($request) {
             $validated = $request->validated();
 
+            $validated['no_regis_app'] = Str::upper(Str::random(12));
+
             if ($request->hasFile('dasar_hukum')) {
                 $dasar_hukumPath = $request->file('dasar_hukum')->store('aplikasi/dasar-hukums', 'public');
                 $validated['dasar_hukum'] = $dasar_hukumPath;
-            }
-            if ($request->hasFile('nda')) {
-                $ndaPath = $request->file('nda')->store('aplikasi/ndas', 'public');
-                $validated['nda'] = $ndaPath;
-            }
-            if ($request->hasFile('sop')) {
-                $sopPath = $request->file('sop')->store('aplikasi/sops', 'public');
-                $validated['sop'] = $sopPath;
-            }
-            if ($request->hasFile('kak')) {
-                $kakPath = $request->file('kak')->store('aplikasi/kaks', 'public');
-                $validated['kak'] = $kakPath;
-            }
-            if ($request->hasFile('capture_frontend')) {
-                $capturefrontendPath = $request->file('capture_frontend')->store('aplikasi/capture-frontends', 'public');
-                $validated['capture_frontend'] = $capturefrontendPath;
-            }
-            if ($request->hasFile('capture_backend')) {
-                $capturebackendPath = $request->file('capture_backend')->store('aplikasi/capture-backends', 'public');
-                $validated['capture_backend'] = $capturebackendPath;
-            }
-            if ($request->hasFile('buku_manual')) {
-                $buku_manualPath = $request->file('buku_manual')->store('aplikasi/buku-manuals', 'public');
-                $validated['buku_manual'] = $buku_manualPath;
-            }
-            if ($request->hasFile('dokumen_perancangan')) {
-                $dokumen_perancanganPath = $request->file('dokumen_perancangan')->store('aplikasi/dokumen-perancangans', 'public');
-                $validated['dokumen_perancangan'] = $dokumen_perancanganPath;
-            }
-            if ($request->hasFile('surat_mohon')) {
-                $surat_mohonPath = $request->file('surat_mohon')->store('aplikasi/surat-mohons', 'public');
-                $validated['surat_mohon'] = $surat_mohonPath;
-            }
-            if ($request->hasFile('implemen_app')) {
-                $implemen_appPath = $request->file('implemen_app')->store('aplikasi/implemen-apps', 'public');
-                $validated['implemen_app'] = $implemen_appPath;
-            }
-            if ($request->hasFile('lapor_pentest')) {
-                $lapor_pentestPath = $request->file('lapor_pentest')->store('aplikasi/lapor-pentests', 'public');
-                $validated['lapor_pentest'] = $lapor_pentestPath;
             }
 
             $newApplication = Application::create($validated);
@@ -130,17 +90,13 @@ class ApplicationController extends Controller
      */
     public function edit(Application $application)
     {
-        $katapps = Katapp::all();
-        $katpenggunas = Katpengguna::all();
-        $katplatforms = Katplatform::all();
-        $katdbs = Katdb::all();
-        $katservers = Katserver::all();
-        $bahasaprograms = Bahasaprogram::all();
-        $frameworkapps = Frameworkapp::all();
-        $layananapps = Layananapp::all();
         $opds = Opd::all();
+        $katpenggunas = Katpengguna::all();
+        $katservers = Katserver::all();
+        $layananapps = Layananapp::all();
+        $katapps = Katapp::whereIn('id', [1, 2])->get();
 
-        return view('aplikasi.aplikasi-edit', compact('application', 'katapps', 'katpenggunas', 'katplatforms', 'katdbs', 'katservers', 'bahasaprograms', 'frameworkapps', 'layananapps', 'opds'), [
+        return view('aplikasi.aplikasi-edit', compact('application', 'opds', 'katpenggunas', 'katservers', 'layananapps', 'katapps'), [
             'title' => 'Data Aplikasi'
         ]);
     }
@@ -153,71 +109,22 @@ class ApplicationController extends Controller
         DB::transaction(function () use ($request, $application) {
             $validated = $request->validated();
 
-            if ($request->hasFile('dasar_hukum')) {
-                Storage::disk('public')->delete($application->dasar_hukum);
+            if (empty($application->no_regis_app)) {
+                $validated['no_regis_app'] = Str::upper(Str::random(12));
+            }
 
-                $dasar_hukumPath = $request->file('dasar_hukum')->store('dasar-hukums', 'public');
+            if ($request->hasFile('dasar_hukum')) {
+
+                if (!empty($application->dasar_hukum)) {
+                    Storage::disk('public')->delete($application->dasar_hukum);
+                }
+
+                $dasar_hukumPath = $request->file('dasar_hukum')->store('aplikasi/dasar-hukums', 'public');
                 $validated['dasar_hukum'] = $dasar_hukumPath;
             }
-            if ($request->hasFile('nda')) {
-                Storage::disk('public')->delete($application->nda);
 
-                $ndaPath = $request->file('nda')->store('ndas', 'public');
-                $validated['nda'] = $ndaPath;
-            }
-            if ($request->hasFile('sop')) {
-                Storage::disk('public')->delete($application->sop);
-
-                $sopPath = $request->file('sop')->store('sops', 'public');
-                $validated['sop'] = $sopPath;
-            }
-            if ($request->hasFile('kak')) {
-                Storage::disk('public')->delete($application->kak);
-
-                $kakPath = $request->file('kak')->store('kaks', 'public');
-                $validated['kak'] = $kakPath;
-            }
-            if ($request->hasFile('capture_frontend')) {
-                Storage::disk('public')->delete($application->capture_frontend);
-
-                $capturefrontendPath = $request->file('capture_frontend')->store('capture-frontends', 'public');
-                $validated['capture_frontend'] = $capturefrontendPath;
-            }
-            if ($request->hasFile('capture_backend')) {
-                Storage::disk('public')->delete($application->capture_backend);
-
-                $capturebackendPath = $request->file('capture_backend')->store('capture-backends', 'public');
-                $validated['capture_backend'] = $capturebackendPath;
-            }
-            if ($request->hasFile('buku_manual')) {
-                Storage::disk('public')->delete($application->buku_manual);
-
-                $buku_manualPath = $request->file('buku_manual')->store('buku-manuals', 'public');
-                $validated['buku_manual'] = $buku_manualPath;
-            }
-            if ($request->hasFile('dokumen_perancangan')) {
-                Storage::disk('public')->delete($application->dokumen_perancangan);
-
-                $dokumen_perancanganPath = $request->file('dokumen_perancangan')->store('dokumen-perancangans', 'public');
-                $validated['dokumen_perancangan'] = $dokumen_perancanganPath;
-            }
-            if ($request->hasFile('surat_mohon')) {
-                Storage::disk('public')->delete($application->surat_mohon);
-
-                $surat_mohonPath = $request->file('surat_mohon')->store('surat-mohons', 'public');
-                $validated['surat_mohon'] = $surat_mohonPath;
-            }
-            if ($request->hasFile('implemen_app')) {
-                Storage::disk('public')->delete($application->implemen_app);
-
-                $implemen_appPath = $request->file('implemen_app')->store('implemen-apps', 'public');
-                $validated['implemen_app'] = $implemen_appPath;
-            }
-            if ($request->hasFile('lapor_pentest')) {
-                Storage::disk('public')->delete($application->lapor_pentest);
-
-                $lapor_pentestPath = $request->file('lapor_pentest')->store('lapor-pentests', 'public');
-                $validated['lapor_pentest'] = $lapor_pentestPath;
+            if (isset($validated['status']) && $validated['status'] === '1') {
+                $validated['alasan_nonaktif'] = null;
             }
 
             $application->update($validated);
